создает модуль генерации фотографий
